Updated server details
