+++ openhands/agent_server/conversation_service.py
@@ -17,6 +17,7 @@ from openhands.agent_server.models import (
     StoredConversation,
 )
 from openhands.agent_server.pub_sub import Subscriber
+from openhands.agent_server.server_details_router import update_last_execution_time
 from openhands.agent_server.utils import utc_now
 from openhands.sdk import EventBase, Message
 from openhands.sdk.conversation.state import AgentExecutionStatus
@@ -264,6 +265,7 @@ class _EventSubscriber(Subscriber):
 
     async def __call__(self, event: EventBase):
         self.service.stored.updated_at = utc_now()
+        update_last_execution_time()
 
 
 @dataclass
